feat: Implement Check Debit Transaction Screen and Location Handling
This commit introduces the Check Debit Transaction screen, allowing users to input transaction details and perform a debit check. It also includes functionality to retrieve the user's current location for inclusion in the transaction data.

- Added `CheckDebitTransactionState` data class to manage the UI state of the screen, including loading state, button enablement, and transaction details.
- Added `CheckDebitTransactionScreen` composable to display the input fields for source account, destination account, destination bank, and memo, along with a button to trigger the debit check.
- Added `CheckDebitTransactionViewModel` to handle the business logic for the screen, including updating the state based on user input and fetching the current location using the device's LocationManager.
- Implemented logic in `CheckDebitTransactionViewModel` to request and update the `geoLocation` state with the user's current location.
- Added `GeoLocation` data class to represent geographical coordinates with latitude, longitude, and accuracy.
- Implemented a check in `CheckDebitTransactionViewModel` to enable the "Check Debit Transaction" button only when all required fields are filled and the screen is not in a loading state.
- Added a loading indicator to the "Check Debit Transaction" button while the debit check is in progress.
- Added permission handling for location access in `DashboardScreen` using `rememberLauncherForActivityResult` and `ActivityResultContracts.RequestMultiplePermissions` to request `ACCESS_FINE_LOCATION` and `ACCESS_COARSE_LOCATION`.
- Modified `DashboardScreen` to request location permissions before navigating to the Check Debit Transaction screen if permissions are not already granted.
- Updated the `GeoLocation` data class to use `Float` for `accuracy`.

+++ app/src/main/java/ng/wimika/moneyguardsdkclient/ui/features/checkdebit/models/GeoLocation.kt
@@ -3,5 +3,5 @@ package ng.wimika.moneyguardsdkclient.ui.features.checkdebit.models
 data class GeoLocation(
     val lat: Double = 0.0,
     val lon: Double = 0.0,
-    val accuracy: Double = 0.0
+    val accuracy: Float = 0.0f
 )
